Add the size to all image names.
Currently, the images names convey that some (non-aftermarket versions
of) computers were originally shipped with a unique flash size, and
that other flash sizes are after-market modifications.

When this happens, the computer images don't have the flash size in
their filenames.

However since some computers were originally shipped with multiple
flash sizes (like the ThinkPad X200 that has both 4MiB for the non-AMT
version and 8MiB for the AMT version), and that the installation
instructions we currently have tell people to identify their flash
size, it is easier to add the flash size everywhere than trying to
explain the situation with computers that have different flash sizes
only due to aftermarket modifications.

This also makes it easier to write better installation instructions.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>