git: gitconfig: don't sign commits.
Without that fix, depending on the host distribution configuration,
gpg could try to sign commits:
    $ make release
    [...]
    HEAD is now at 56ccc5ed5 util/grub.d/30_os-prober.in: Fix GRUB_OS_PROBER_SKIP_LIST for non-EFI
    Applying: mitigate grub's missing characters for borders/arrow characters
    error: gpg failed to sign the data
    fatal: failed to write commit object
    make: *** [Makefile:840: release] Error 128

In the case above, this is a Trisquel 11 x86_64 installation that is
at the same time used as a chroot (from within Guix system
x86_64-linux) and that can also be booted, which results in a git
and/or gpg host configuration made to sign commits, while failing when
in a chroot.

Fixing that issue also improves reproducibility as the host
configuration should not impact how GNU Boot's git commands import
patches.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>