manual: Document how to use GNU Boot with QEMU.
Before the only documented way to use QEMU to run GNU Boot was to
"look in the GNU Boot source code as GNU Boot uses QEMU to run some
automatic tests that boots Trisquel 11 (aramo)." which in practice
means that users had to find these and then adapt the commands to boot
another image than the one produced by the GNU Boot automatic tests.

Users could also ignore that advice and directly try to get something
working by trial and error. The problem is that this is also hard as
many parameters need to be right:

- The machine need to be set to pc as so far we only provide images
  for pc and not q35.

- To make SATA emulation fast and/or to make sure that the boot
  doesn't take too long, we need a lot more parameters than just -hda
  <path/to/image>.

  We also don't use qcow2 on purpose as it minimize the difference
  with a real installation.

  By default QEMU autodetects the file type, and the VM should boot
  even with raw images, but when that happens it prints the following
  warning:

      WARNING: Image format was not specified for 'Makefile' and probing
               guessed raw.
               Automatically detecting the format is dangerous for raw
               images, write operations on block 0 will be restricted.
               Specify the 'raw' format explicitly to remove the
               restrictions.

  Since we also want to give users fully working commands, we cannot
  use -hda and have to instead tell QEMU that we are using a raw
  image.

- The RAM size also need to be adjusted to match the image of the
  distribution while at the same time not eating too much RAM in case
  people don't have a lot of RAM. This can happen if people got a
  compatible computer but didn't add more RAM yet. So here providing
  known good images can help making sure the RAM usage is minimal.

So now users also have the option to just follow instructions to get
something working. And since we don't hide the QEMU commands, users
also have an easier starting point to boot the distribution of their
choice.

In addition to all that, users also need to handle the case where
/dev/kvm doesn't have the right permissions. Skipping that would make
the boot very slow. The loss of speed is somewhat acceptable on
extremely minimalist distributions images without Xorg or Wayland
and/or a kernel and an initramfs.

However we also want empower users to run the distributions they wish,
and for instance the default Trisquel installer comes with a full
blown desktop environment, and if we didn't make users adjust the
/dev/kvm permissions, it would make running distributions like
Trisquel much harder.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>