packages: grub: download: is_done: fix patches detection.
The 'for grubpatch in ../resources/grub/patches/*; do' runs from the top
directory and therefor can't work.

It was copied from the download function, but there these commands are run
inside the grub (temporary) directory.

In addition 'if [ "${current_commit}" != "${grub_revision}" ]' was always
false since we apply patches to GRUB.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>