Add GNU Boot logos sources and use them to generate the GRUB backgrounds.
This adds the original logo and background sources in order to be able
to reuse them for various purposes (for instance making favicons,
etc).

In addition the GRUB background are now generated with code from one
of these source logos because it makes it easier to modify them and/or
to generate additional background for additional resolutions.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>