website: status: remove unused installation instructions entries.
In the commit 540bcca0fbc94e6a051e1620f126875d699df7a7 ("website:
status: Add GNU Boot 0.1 RC5."), In the legend for the status, the
"Installation instructions" part of this legend was accidentally
removed from the English status page.

However we never got any report of tests of the "Installation
instructions" or "Upgrade instructions", so it is better to remove
these entries from the tables for the previous RC releases.

Since this wasn't removed from the Spanish status, we also need to
remove it there for consistency.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>
Tested-by: Adrien 'neox' Bourmault <neox@gnu.org>