Makefile.am: Make sure all errors are written to logs.
While there are already some 2>&1 to also write errors to logs, not all
commands had them.

So for instance when running 'make release', some of the errors are not
written to logs.

The 2>&1 | tee [...] were also aligned together to be able to check more
easily if some 2>&1 | tee [...] were missing or misspelled.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien Bourmault <neox@gnu.org>