Add --with-xz-extra-args.
When tracking the following regression (with new patches):

    [ PASS ] check_cbfs_image_log: found all known required files [...]
    guix build: warning: source expression failed to match any pattern
    guix build: error: gnuboot-trisquel-grub.img: unknown package
    cp: cannot stat '/share/trisquel-installer/grub.img': No such file or directory
    make: *** [Makefile:909: release] Error 1

xz was taking a lot of time, even on a powerful computer (a VM with 8 cores
and 16 GiB of RAM that runs on a KGPE-D16).

The xz manual says that "The single-threaded and multi-threaded compressors
produce different output.". It however doesn't say if a different number of
threads can produce different output in multi-threaded mode, so for now we
keep it single-threaded by default instead of enabling multi-threaded
compression-decompression with only 1 thread by default.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Tested-by: Adrien Bourmault <neox@gnu.org>
Acked-by: Adrien Bourmault <neox@gnu.org>