website: manual: Fix picture display in the downloaded HTML version.
The picture(s) are supposed to be in images/, so downloading the
picture in the same directory than the manual page won't work.

However when we use "Save Page As ...", in Firefox and derivatives,
the picture works as it takes care of including all the pictures and
necessary files inside a subdirectory.

The big advantage of this new method is that it also scales well if we
increase the number of pictures.

This was broken from the start in the commit
08b9e449e94e99c0ec924e2026dc650b63ddc3e5 ("Add a minimal GNU Boot
manual.").

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Tested-by: Adrien Bourmault <neox@gnu.org>
Acked-by: Adrien Bourmault <neox@gnu.org>