manual: align nodes and require it too.
This makes it easier to check if the node name matches the
chapter/*section.

A test to check for false positive was not added as there is no commit
yet in the tree that has aligned nodes. This test will need to be
added after commits with aligned nodes are pushed in the main branch
to avoid potential mistakes with commit hashes.

The tests/checkpatch code was tested to always print something when a
test fail, by inserting check rules that fail on specific commits.

For instance the following check rule was added (temporarily):
    (make-rule
        "Crash checkpatch.scm for a test"
        (lambda (check-results) check-results)
        (lambda (line parse-results check-results) #t)
        (lambda (line parse-results check-results) check-results)
        (lambda (path parse-results check-results)
          (if (string=? (basename path)
    		    "0001-manual-Describe-the-GNU-Boot-project.patch")
    	  crash)
          check-results))

to test if the following code:
    # This checks the "manual: Describe the GNU Boot project." commit
    # which has non-modified unaligned nodes.
    ./scripts/checkpatch.scm \
        "$(git format-patch -1 5303fab6a49796d69fc9176b715409f925d8f608)" > \
        "${tmpdir}"/checkpatch-test-non-modified-unaligned-nodes.log \
        2>/dev/null || \
        fail "checkpatch.scm failed " \
             "when checking non-modified unaligned nodes in the manual."

would really print "checkpatch.scm failed when checking non-modified
unaligned nodes in the manual." when checkpatch.scm crashes or exits
with an error.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>