checkpatch.scm: check if the patch commit hash is in git.
We want to implement this requirement so that we can later rely on
Git to more easily and reliably determine if a file exists at the
commit hash of a specific patch. For this to work, the patch must
have a valid Git commit hash and the hash needs to be in git as well.

Checking for file existence can be required to check if translations
were updated, as there is no point in adding a warning if the file
doesn't exist once the commit is applied, and checking current files
with older commit also doesn't make any sense.

As for the implementation, the git command was used directly to make
it easier for contributor to run the tool as guile-git might not be
packaged everywhere. And while git-lib is a dependency of Guix, we
also want to use checkpatch.scm for checking patches touching the
website, and we don't want to require the use of Guix, or a
distribution that has Guix or guile-git for patches related to the
website.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
neox: improved commit message and fixed typos
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>