checkpatch.scm: check if the patch commit hash is in git.
We want to require this to be able to later on depend on git to be
able to (much more easily and reliably) find out if a file exist at
the commit hash of a certain patch. To do that the patch needs to have
a valid git commit hash and the hash needs to be in git as well.

Checking for file existence can be required to check if translations
were updated, as there is no point in adding a warning if the file
doesn't exist once the commit is applied, and checking current files
with older commit also don't make any sense.

As for the implementation, the git command was used directly to make
it easier for contributor to run the tool as guile-git might not be
packaged everywhere. And while git-lib is a dependency of Guix, we
also want to use checkpatch.scm for checking patches touching the
website, and we don't want to require the use of Guix, or a
distribution that has Guix or guile-git for patches related to the
website.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>