checkpatch.scm: warn about missing translations.
We often need to update the status table. And when we do that,
sometimes we only change the value of a given field, and the new value
is also used somewhere else.

For instance we might change this line:
    +------------------+--------------+--------------+
    | Lenovo,          | Tested       | Untested     | [...]
    | ThinkPad X200    |              |              |
    +------------------+--------------+--------------+
to become this line:
    +------------------+--------------+--------------+
    | Lenovo,          | Tested       | Tested       | [...]
    | ThinkPad X200    |              |              |
    +------------------+--------------+--------------+

In a case like this one, if we want, we can require the patch to also
update the translations at the same time, as knowledge of the language
or even character set is not required to do such work.

However that work can be tedious to do in the case where the
contributors are not used to the character set. In such case the
contributors might need help for doing that work.

Since maintainers are supposed to run checkpatch.scm on contributor
patches, this should make them remember to check for situations like
the one described above, and since it is only a very vague warning, it
won't alarm users of checkpatch.scm too much, and this should give the
maintainers the freedom to deal with this situation as they see fit.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>