Add Debian package for the GNU Boot manual.
This enables to more easily test the manual build system for people
running Trisquel 11 (aramo).

Adding files in debian/ conflicts with Debian's attempt to package
things.

Since Linux can also produce Debian packages, we used the same
Makefile target name ('debian') and instead of creating files in
debian/ they were created in resources/debian and just copied in the
right location.

This was tested on Trisquel 11 (aramo) with Emacs, and the SOIC-16
picture was displayed.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>