website: get rid of web/ prefix.
Since the commit be72474aa04f1de30fc1755af41094e2fb969dac
("website-build: Use rsync for deploying the website."), we use rsync
instead of CVS for deployment (especially because CVS was too complex
to deal with for us because of its speed and limitations).

This makes the migration to haunt easier.

In addition we will most likely never write a tool to make deploying
websites to CVS fast and move back to CVS.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>