website: markdown: switch to Haunt compatible metadata.
While it is relatively easy to write code for Haunt to parse the
metadata of the markdown files we currently have, and that my previous
attempts to migrate to Haunt did exactly that, not doing that reduces
a lot the amount of code required for the migration.

In addition it also enables to split more the big patch that migrates
to Haunt.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>