[WIP] website: Add support for building with Haunt.
TODO:
- Add support for Spanish
- Adjust generated HTML tables with a number of column < 5.

This add support for Haunt as an additional website generator, and the default
remains Untitled.

In addition this change also brings the ability to compare the output of both
Haunt and Untitled to be able to then decide if or when to migrate to Haunt
and remove support for Untitled completely.

This can be done with the following commands:
    $ cd website
    $ ./autogen.sh
    $ ./configure
    $ make lbssg-website-pretty-html.tar haunt-website-pretty-html.tar
    $ mkdir old new
    $ tar xf lbssg-website-pretty-html.tar -C old
    $ tar xf haunt-website-pretty-html.tar -C new
    $ meld old new

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>