website: remove template.original.
This file was kept to show what was modified from it to create
template.include, however we don't need this kind of information as we
are migrating to Haunt and we will rewrite the templates along the
way.

In addition the authorship is already preserved and this file will
remain forever in the git history.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>