checkpatch.scm: enable line and line-match rules to access the path.
We are parsing and checking a patch, so we don't really need access to the
patch path in these functions.

However extending checkpatch.scm to also be able to check files will require
these functions to access file paths.

This could for instance allow to skip checks in the line-match function if the
path doesn't have a specific file extension.

It could also allow to match a file name with the declared guile module name
in the line function.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>