[WIP] website: use recfiles to generate status pages.
TODO:
- Make it possible to not bail out if AC_PROG_LOCALE is not found but set
  variable that then in Makefile.am conditionally build the locale
- wrap the code that uses locales with Guix when not using --disable-guix.
- Handle the resources/pakcages/guix (generate locales locally?)
- make cmp work
- checksum [1]

This generates the exact same file than website/pages/status.md and
website/pages/status.es.md and it automatically checks that (with cmp)
during the build of the website.

This is done to make it easier to review as the generated status file is
exactly the same.

The main goal of this change is to make it easier to update the status
tables, especially in the context of the migration to Haunt that makes
finding a solution even more necessary.

Even if we don't migrate to Haunt, this modification was badly needed:
While working on this patch, multiple errors were found in the status
tables (and fixed in earlier commits). This shows that mistake could
very easily slip through if we need to update the tables by hand.

As it is now, even if the build will fail when something is changed,
it is already possible to do more easily update the tables: since the
status is generated before checking, the non-generated status can
easily be updated with meld once the check fails.

As for the migration to Haunt, we plan to keep markdown files for the
website, but Haunt only support the CommonMark variant of Markdown,
which is a standardization effort for Markdown.

CommonMark didn't define a syntax for tables, and in theory this
shouldn't be a big issue since the CommonMark standard also allows
valid HTML, and so in theory it should be possible to create tables
with HTML.

However guile-commonmark doesn't support HTML, so a pull request[1] to
add support for HTML was sent to it in October 2024, but so far it
wasn't reviewed yet, so guile-commonmark is probably not maintained
anymore.

[1]https://github.com/OrangeShark/guile-commonmark/pull/23

Many workarounds are possible. For instance it is also possible to use
plain HTML pages within Haunt, or to generate pages with scheme. There
is also a fork with HTML support by The Spritely Institute, but it
didn't reach distributions like Guix yet (though this could also be
fixed).

While the HTML syntax is probably good enough for technical users, as
the syntax enables to not have ultra long lines, any of these
solutions would still make it hard to update the status for less
technical users.

In addition, parsing and/or reusing the recfiles is relatively easy as
Guix has code to do that, and the GNU recutils can also be used to
easily retrieve data from recfiles from within shell scripts or even
Makefiles.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>