manual: move internal SD card readers to new computers subsections.
We want to have computers subsection to be able to document very
specific hardware, like which boot flash variants exist for a given
computer, or which CPUs are available, what is removable or soldered,
what is not, etc.

However internal card readers are currently in a section about
peripherals makes which then makes it complicated to understand where
to add information about hardware support.

So this moves the SD card readers information in computers
subsections. But if these section would only have some information
about microSD readers, they would look a bit strange. So a bit more
information was added along the way such as the computer vendor,
product, max RAM, the CPU available, etc.

Since we are also adding information like the boot flash size, we also
need to clarify that the entry about the "ThinkPad X200" isn't valid
for an X200 sold by Minifree Ltd.

This decision is documented in the commit
f148812d19018396580a0516680147c63d80ca74 ("manual: Add list of
compatible computers."), but readers of the manual are not expected to
read all the GNU Boot commits before starting to read the the manual,
and without a clear explanation of this decision, it could make things
a bit confusing, especially when readers have a computer sold by
Minifree Ltd, which isn't in the list yet (we also want to add them
but we would need more information on these computers to do it well).

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>