website: status: 0.1-RC6: mention how the release was built.
This helps with transparency but also reproducibility.

With this information, users can try to use Trisquel 11 (aramo) over
PureOS 10 (byzantium) to limit the risk of non-working images.

It also gives a bit the situation in which the images were built to
help evaluate the risk of using the released images instead of using
self-built images.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>