dependencies: arch: update copyrights.
The copyrights were obtained with the following command, in the GNU
Boot repository but also in the libre branch of osbmk, and in the tag
r20160907 of the "obsolete repository preserved for historical
purposes": 'git log --oneline --pretty="# Copyright (C) %ad %an <%ae>"
--follow --date="format:%Y" -- PATH".

Path was replaced by the path of the file, which depends on the
repository.

Before the first commit (in the first release tarballs), all the
copyrights were 2014 from Leah.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>