resources: dependencies: sync arch and pureos-trisquel.
This makes sure that arch/Parabola have the same dependencies than
Trisquel.

While libselinux is packaged in Parabola, most SELinux related
libraries were only added to be able to compile foreign distributions
like Replicant, and SELinux is not integrated in Parabola (for
instance there are no SELinux policies at all). Because of that
libselinux was not added in the 'arch' file.

It was kept in pureos-trisquel because it was introduced in the README
of the very first Libreboot tarball release[1] and there is so far no
documentation on why it is really needed. It could be removed if
builds are found to work well without it.

[1]https://rsync.libreboot.org/oldstable/20131212/X60_source.tar.gz

The copyright headers were also synchronized as we are merging code
back and forth between the two files. The old mails were updated and I
also updated my copyright along the way.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>