website: warning banner: warn against (potentially) unsafe infos.
pages/docs/install/spi.md has:

    Use quality rosin flux (not acid based) and good 60/40 or 63/37
    leaded solder (don't use lead-free)

This lacks advices to wash hands to prevent potential lead
poisoning.

We have that instead:

    In case you're not comfortable with soldering, we have some
    excellent videos linked on the [FAQ page](../../faq.md) which you
    can watch.

And the FAQ then has a section about "Where can I learn more about
electronics" (with good references but no direct mention of safety
issues).

In contrast soldering courses have advices like that[1]:

    Lead is poisonous. It gets on your skin when you hold the
    solder, so wash your hands after soldering!

While most people who do soldering know that, we still need to remind
people of in case they missed it.

The issue is also present for more advanced topics. We have this in
the same file:

    Make sure to buy 99.9% isopropyl alcohol. Don't buy weaker
    solutions because they contain water, and don't use other
    chemicals because most other chemicals are corrosive. You use the
    isopropyl to clean the area you're soldering, before soldering it,
    and then soak up the wet alcohol with a cloth. You will also use
    it to clean off any flux that you used.

According to Wikipedia[2], Isopropyl is Flammable, and mildly toxic,
but there are no advices on how to handle it or to try to not breath
it.

So this means that these articles are meant for people (self) trained
in electronics.

Since GNU Boot is meant for a wider audience and advertized as such we
can't skip the proper warnings.

Here it is also best to warn instead of trying to fix everything
because many pages have not been properly reviewed.

[1]From "Soldering is easy here's how to do it", under the CC-BY
   license, available at https://mightyohm.com/soldercomic
[2]https://en.wikipedia.org/wiki/Isopropyl_alcohol#cite_ref-iupac1979_1-0

To properly show that we have multiple issues, we added bullet point
work in the banner by removing "header ul>li". However this broke many
things:

- The menu became a vertical list, so we replaced it with ".menu ul>li"
  instead to get it back like it was.

- It broke the display of the anchor links, so to fix it the ul>li
  below was removed.

- It increased margins a lot around each line, the rest of the CSS changes
  are to fix that.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>