website: warning banner: warn against breaking computers.
In website/pages/docs/install/t400_external.md we have:

    An “HMM” (Hardware Maintenance Manual) detailing the process of
    [dis]assembly is available for this model. Be careful when
    reassembling the laptop as the screws on page 114 (with title
    “1130 Keyboard bezel”) are swapped and if you follow the HMM you
    will punch a hole through the bezel in the upper right corner.

But "Flashing the ThinkPad R400" (in
website/pages/docs/install/r400_external.md) doesn't advise to remove
the battery before disassembling the computer.

While that might look like common sense for people with some
electronic knowledge, there are also advises that require more
computers skills to follow properly without the risk of breaking the
computers.

For instance in pages/docs/install/spi.md we have:

    After you're done, use a soft bristle brush and 99.9% isopropyl
    alcohol to break up the remaining flux, then soak up the flux
    using a cloth, while the alcohol is still wet. 99.9% isopropyl is
    the best liquid to use, because it evaporates quickly and it does
    not leave a corrosive residue.

According to neox (who is trained in electronics) while 99.9%
isopropyl doesn't provoke corrosion directly, it can do it indirectly
because of its high polarity on one extremity and weak polarity on the
other one (carbon chain), making it a radical, and according to neox
this is what makes isopropyl destroy plastics (like chips casing) and
even circuit protection coatings (dissolving them) through hydrolyse.

So, according to neox, the way to prevent that is to not let it on the
circuit for a long time by using a small tool to apply and let it dry
quickly.

While all that might be obvious for someone trained in
electronics (it's part of the curriculum) or electronics repairs (it
would be hard not to know about it since isopropyl is widely used to
remove corrosive and/or conductive flux), it is not obvious for
everybody, including people with basic soldering (self) training,
hence the warning.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
neox: fixed typos in commit message
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>