website: get rid of web/ prefix.
Since the commit be72474aa04f1de30fc1755af41094e2fb969dac
("website-build: Use rsync for deploying the website."), we use rsync
instead of CVS for deployment (especially because CVS was too complex
to deal with for us because of its speed and limitations).

This makes the migration to haunt easier.

In addition we will most likely never write a tool to make deploying
websites to CVS fast and move back to CVS.

The old links should continue to work as we added a .symlinks[1]
file. Thanks for Jason Self for the help on how to deal with
redirections and also with help that lead me to find the right
documentation.

The same documentation mentions "Server-side scripts"[2], and so the
symlinks script comes from there.

[1]https://www.gnu.org/server/standards/gnu-website-guidelines.html#symlinks
[2]https://www.gnu.org/server/standards/gnu-website-guidelines.html#scripts

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>