grub.cfg: implement local variables scope.
Local variables are emulated by making them use use globally unique
names that also have the function name inside.

This avoids variable names collisions, and enables to easily
check (when reading the code) if variables are set in the right
functions. Function names also need to be globally unique.

To get the shortest names possible, these look like this:
    _<function name>_<variable name>
but as this increase some line lengths a lot, some function names or
variable names were renamed to be shorter.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>