website: move "support for more computers" from FAQ to contribute.
This is something that is directly relevant to people wanting to
contribute to GNU Boot, so it makes more sense to have it there.

In addition the FAQ still has a scary warning, so moving
information (to a better place) also shows that this information not
affected by any of the issues described in the warning.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>