manual: move security in its own chapter.
Security need to be put back in context. To do that we bring the
notions of risk management and threat modeling.

This help users avoid pointless efforts of increasing security just
for the sake of security. In contrast we also need to warn users
against the really bad consequences of security breaches, even if they
are really rare when using only free software.

Trying to explain all that in all sections that will have security
impact, like installation instructions, flash programmers, PCIe cards,
etc, is probably a bad idea because it will only enable to talk about
specific threats without putting them back in a bigger context.

And we also can't assume that all users already use risk management
and/or threat modeling tailored to their use of computers, because
there is a lot of propaganda (within the computer industry, but also
some that target a wider audience) about security that assume
Enterprise contexts or that users don't run 100% free software.

For instance not writing password on paper is a good practice in a
company office, but not doing that is often advertized without any
context, however it can actually be very safe for a person to write
passwords on a paper notebook and to have that person hide that
notebook her home when the probability of someone else going into that
home and finding that notebook is extremely low.

Another example would be the pressure to implement some form of
signature verification for the boot chain, which in a lot of cases
ends up being restricted boot. This is also being advocated by members
of the FOSS community, including people with advanced technical
knowledge without putting all that in context.

If we stopped there, we would end up with the "Security features"
section being outside of the Security chapter, which isn't logical, so
to make everything fit we also include a "Security considerations when
using GNU Boot" section, and we moved the information about secure
boot and restricted boot under a subsection of that section.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>