build, download: Handle unusable /dev/kvm.
So far non-usable /dev/kvm was not handled in ./build or ./download
scripts.

Not being able to use /dev/kvm brings in serious performance
degradation, so in that case we cannot silently disable KVM usage.

So here too need users to take action and decide what is better for
them (between not using KVM or giving the current user permission to
use /dev/kvm).

Bug-report: https://savannah.gnu.org/bugs/?66225
Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>