website: menu: remove FAQ.
The FAQ has a big scary banner. Given the amount of information in the
FAQ is quite big (1208 lines), and fixing some of the entries require
significant work.

For instance for fixing 'Do not use CH341A!', we needs more
information on which flash programmers do have 3.3v and which ones
have 5v, especially because some hardware vendors that also sell RYF
certified hardware sell some, so we might be able to provide advices
that avoid mistakes and that are more fine grained.

Some information may or may not be outdated and need more testing. For
instance 'The ethernet doesn’t work on my X200/T400/X60/T60 when I
plug in it', 'My KCMA-D8 or KGPE-D16 doesn’t boot with the PIKE2008
module installed', etc.

Fixing all that can require a lot of work, and time might be better
spent in other areas of GNU Boot.

A strategy that worked so far was to move things away from the FAQ to
pages that didn't have this banner, for instance in the commit
d702c966f693b19d3dc87c7882a1fda4300ebdc4 ("website: move "support for
more computers" from FAQ to contribute."), or to document specific
flash chip programmers like the Vikings CH341a USB Programmer in the
GNU Boot manual.

Given the issues and the strategy with the FAQ it doesn't really make
sense to highlight the FAQ in such prominent way.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>