website: downloads: remove unreviewed banner.
None of the information in the "Statically linked" section is relevant
anymore for GNU Boot: for building i945 flashing utilities, we use
Guix and not Trisquel, and we simply provide all the source code with
Guix, and it is released at the same place than the other binaries.

Since nothing is special about any of that, it doesn't need to be
mentioned in Downloads.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>