[WIP] website: docs: build: debootstrap: replace table by generic text.
The table we had and the text that introduce it both required
maintenance.

The table listed host and target distributions, and it was never
updated for Trisquel 11.

Since the commit 17fb225622e79cb16eb2c843ee49ac62bb610113 ("website:
docs: build: advise to use the same distro than the release."), the
text that explained the table also became outdated and less clear: it
only mentions PureOS 10 or Trisquel 10, while the preferred
distribution to use GNU Boot is now Trisquel 11.

So to reduce maintenance, the table was replaced by a more generic
text that works for all distributions and also explain the risks of
using debootstrap without the appropriate keyring packages.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>