website: history: replace table.
We want to migrate to Haunt. Haunt support various formats. It also
supports the CommonMark variant of Markdown, so the plan is to reuse
our Markdown files as much as possible.

The only way to use tables in CommonMark is to use HTML to define
tables. Unfortunately Haunt depends on guile-commonmark which has an
incomplete CommonMark implementation as it doesn't support HTML.

There is a patch for guile-commonmark that adds HTML support, and this
patch is being used for the website of the spritley institute, however
guile-commonmark is also unmaintained as the patch has been waiting
for a review for more than one year.

In order to keep things simple and not to have to recompile
guile-commonmark and Haunt, the plan is to only use tables when this
is strictly necessary, and to generate the pages that use tables
through other means (like scheme code for instance).

When creating tables is necessary, the idea is also to take advantage
of that and generate the table from data in a format that is easy to
edit (like recfiles) and that can also be reused (in scheme, shell
scripts, etc).

Beside making it possible to migrate to Haunt, this change also makes
the information fit within 80 characters lines. It also makes it
easier to edit the information, though this is probably not very
useful here as this information will probably not change a lot.

The downside is that the information is now less compact.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>