website: status: Upstream versions: reduce maintenance.
Without this change, we need to update the section at every release, just to
update the revision of the latest GNU Boot release, and we already missed 3
releases (0.1 {RC4,RC5,RC6}).

Even with this change, we can still forget to update this section.

However the upstream version section is right for a bigger amount of time
without maintenance: instead of having to update the section at each releases,
we now only need to do it at the release after changing an upstream revision,
and we only need to do that if the upstream revision is based on a new
upstream release.

In addition it is less hard to forget about updating this section since the
release notes might contain information about updated upstream versions, and
we might also not forget to update the upstream project revisions, and when we
do that the "Since GNU Boot release" column will very hard to miss.

Since the table was rearanged a bit I also took the occasion to fixup the text
arrangement to make it more consistent with the other tables. This consistency
also enables to more easily generate the table content automatically.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>