website: status: English: Limitations: break the table.
We want to migrate to Haunt. Haunt support various formats. It also supports
the CommonMark variant of Markdown, so the plan is to reuse our Markdown files
as much as possible.

The only way to use tables in CommonMark is to use HTML to define
tables. Unfortunately Haunt depends on guile-commonmark which has an
incomplete CommonMark implementation as it doesn't support HTML.

There is a patch for guile-commonmark that adds HTML support, and this patch
is being used for the website of the spritley institute, however
guile-commonmark is also unmaintained as the patch has been waiting for a
review for more than one year.

In order to keep things simple and not to have to recompile guile-commonmark
and Haunt, the plan is to only use tables when this is strictly necessary, and
to generate the pages that use tables through other means (like scheme code
for instance).

But here, using or even generating a table here is not a good idea in the
first place. The proof of that is that replacing the table with a list enable
to add more details, and to make things easier to understand.

It also fits the line within 80 characters.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>