manual: add warnings for SSI-EEB mainboard(s).
I was lacking the I/O plate for the connectors of the KGPE-D16 I
used. So after getting one, I added it and also removed some
peripherals to reproduce a Guix configuration that was to be used
later on to reproduce a GPU issue with an Nvidia GPU.

At the end the computer didn't boot anymore and I had the following
issue which is typical of a broken mainboard:

    AMD_CB_EventNotify: HARDWARE FAULT: HT_EVENT_HW_SYNCHFLOOD

I was more careful than with ATX mainboard but for some reasons that
wasn't enough and the RAM init failed, at some point the mainboard
broke. I didn't use an antistatic wrist strap, and I only touched the
computer case to discharge potential static electricity before
operating inside the case. However the computer case was not always
connected to ground as the computer also needs to be disconnected from
electricity when doing such operations.

Given that mainboard with these hardware errors are pretty common, it
is a good idea to be even more cautious.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>